refactor(coordinator): move ForcedTransactionsApp instantiation to CoordinatorApp
Extract ForcedTransactionsApp creation from ConflationAppV1's constructor default into
CoordinatorApp so both ConflationAppV1 and ConflationAppV2 share a single instance.
Move the FTX app lifecycle (start/stop) to CoordinatorApp so it outlives V1 during the
RISC-V transition. Co-locate the proverClientFactory, l2EthClientForConflation,
zkStateClient, and tracesClients in CoordinatorApp to avoid duplicate client instances.
Wire the shared ForcedTransactionsApp into ConflationAppV2 for future safe-block-number use.
